+++ modules/main/src/main/java/com/nisovin/shopkeepers/commands/shopkeepers/CommandSetCurrency.java
@@ -1,9 +1,13 @@
 package com.nisovin.shopkeepers.commands.shopkeepers;
 
+import java.util.ArrayList;
+
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 
+import com.nisovin.shopkeepers.api.ShopkeepersAPI;
 import com.nisovin.shopkeepers.api.ShopkeepersPlugin;
+import com.nisovin.shopkeepers.api.shopkeeper.player.PlayerShopkeeper;
 import com.nisovin.shopkeepers.commands.lib.CommandException;
 import com.nisovin.shopkeepers.commands.lib.CommandInput;
 import com.nisovin.shopkeepers.commands.lib.arguments.StringArgument;
@@ -63,15 +67,26 @@ class CommandSetCurrency extends PlayerCommand {
 			return;
 		}
 
+		// Close all open player shopkeeper UIs to ensure that the new currency item is immediately
+		// in effect everywhere.
+		// Note: This will not save any currently edited trades of open editors.
+		// TODO Inform players why their UI sessions have been closed.
+		// TODO Limit this to player shopkeepers that are actually affected by the currency item
+		// change. Maybe let the shopkeepers themselves react to the settings change.
+		new ArrayList<>(ShopkeepersAPI.getUIRegistry().getUISessions()).forEach(uiSession -> {
+			if (uiSession.getShopkeeper() instanceof PlayerShopkeeper) {
+				uiSession.abort();
+			}
+		});
+
 		if (baseCurrency) {
 			Settings.currencyItem = new ItemData(newCurrencyItem);
-			Settings.onSettingsChanged();
-			Settings.saveConfig();
 		} else {
 			Settings.highCurrencyItem = new ItemData(newCurrencyItem);
-			Settings.onSettingsChanged();
-			Settings.saveConfig();
 		}
+
+		Settings.onSettingsChanged();
+		Settings.saveConfig();
 
 		TextUtils.sendMessage(player, Messages.currencyItemSetToMainHandItem,
 				"currency", currency.getDisplayName(),